requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
Werkzeug==3.0.1
gunicorn==21.2.0
waitress==3.0.0
cryptography==41.0.7
itsdangerous==2.1.2
beautifulsoup4==4.12.2
